gcp/esame-23-febbraio-2023/src/requirements.txt
flask
flask-restful
WTForms
requests
google-cloud-firestore
google-cloud-pubsub
gunicorn
python-dateutil